Move PainlessExecuteAction into action package
